Appease Clippy

+++ yggdrasil/src/filter/audio_input.rs
@@ -74,7 +74,7 @@ fn microphone_input(device: Arc<Mutex<PCM>>) -> Result<AudioSample> {
     let io_device = device.lock().expect("Failed to lock device.");
     let io = io_device.io_f32().into_diagnostic().expect("Failed to io.");
 
-    let mut interleaved_buffer = vec![0.0 as f32; NUMBER_OF_SAMPLES * NUMBER_OF_CHANNELS];
+    let mut interleaved_buffer = vec![0.0_f32; NUMBER_OF_SAMPLES * NUMBER_OF_CHANNELS];
     let number_of_frames = io.readi(&mut interleaved_buffer).into_diagnostic()?;
 
     assert_eq!(number_of_frames, NUMBER_OF_SAMPLES);
